requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
sqlalchemy==2.0.23
werkzeug==3.0.1
jinja2==3.1.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.25.2
pypdf2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
wordcloud==1.9.2
spacy==3.7.2
nltk==3.8.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
pillow==10.1.0
bcrypt==4.1.2
email-validator==2.1.0
langdetect==1.0.9
scikit-learn==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
oauthlib==3.2.2
pyjwt==2.8.0